daemon: fix restoring container with missing task
Before 4bafaa00aa810dd17fde13e563def08f96fffc31, if the daemon was
killed while a container was running and the container shim is killed
before the daemon is restarted, such as if the host system is
hard-rebooted, the daemon would restore the container to the stopped
state and set the exit code to 255. The aforementioned commit introduced
a regression where the container's exit code would instead be set to 0.
Fix the regression so that the exit code is once against set to 255 on
restore.

Signed-off-by: Cory Snider <csnider@mirantis.com>
(cherry picked from commit 165dfd6c3eada774276a4db0038cddd9a28bc9a8)
Signed-off-by: Sebastiaan van Stijn <github@gone.nl>