[20.10] Backport Restore active mount counts on live-restore
Backporting commit 647c2a6cdd86d79230df1bf690d0b6a2930d6db2 for 20.10

When live-restoring a container the volume driver needs be notified that
there is an active mount for the volume.
Before this change the count is zero until the container stops and the
uint64 overflows pretty much making it so the volume can never be
removed until another daemon restart.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>