api: Add a proper deprecation warning to DefaultNetworkSettings
This struct is only used to report the networking state of the default
bridge when inspecting a container.

It was deprecated in v1.09 (API v1.21) but has never been removed since
then. Unfortunately, the deprecation warning was wrongly formatted,
thus automation tools and IDEs are not flagging it as deprecated.

This commit fixes the deprecation comment.

Signed-off-by: Albin Kerouanton <albinker@gmail.com>