Fsync layer metadata on creation
Fsync the link and lower layer files immediately after writing to ensure
that their content is stored on persistent storage. This helps
to address the issue that arises when image layer downloads or
extractions are interrupted due to a power cut.

The Docker daemon can recover if the interrupted layer is "broken,"
but it cannot recover if two or more lower layers to the interrupted
layers are "broken." The suggested fix ensures that the lower layers,
which have been fully downloaded and extracted, remain consistent
even if the current top layer is broken.

Signed-off-by: Mike <mike.sul@foundries.io>