client: ServiceCreate, ServiceUpdate: omit "version" header on API >= 1.30
The "version" header was added in c0afd9c873183604e57282e1cf47605c1f1e4d43,
but used the wrong information to get the API version.

This issue was fixed in a9d20916c3d1f5e3c3ddada79af19aee4ec64629, which switched
the API handler code to get the API version from the context. That change is part
of Docker Engine 20.10 (API v1.30 and up)

This patch updates the code to only set the header on APi v1.29 and older, as it's
not used by newer API versions.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>