c8d: ImageManifest now keeps the descriptor of the manifest
The ImageManifest used to store the manifest of the parent
index/manifest list, this means that users would see multiple same
entries in the images list when a multi-platform image is pulled/built.

Using the manifest instead makes is so that we now see different entires
for each platform image.

Signed-off-by: Djordje Lukic <djordje.lukic@docker.com>