quota: remove gotest.tools from testhelpers
gotest.tools has an init() which registers a '-update' flag;
https://github.com/gotestyourself/gotest.tools/blob/a80f057529047c44e1a85d0d017b200787e537e0/internal/source/update.go#L21-L23

The quota helper contains a testhelpers file, which is meant for usage
in (integration) tests, but as it's in the same pacakge as production
code, would also trigger the gotest.tools init.

This patch removes the gotest.tools code from this file.

Before this patch:

    $ (exec -a libnetwork-setkey "$(which dockerd)" -help)
    Usage of libnetwork-setkey:
      -exec-root string
            docker exec root (default "/run/docker")
      -update
            update golden values

With this patch applied:

    $ (exec -a libnetwork-setkey "$(which dockerd)" -help)
    Usage of libnetwork-setkey:
      -exec-root string
            docker exec root (default "/run/docker")

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>
(cherry picked from commit 1aa17222e7a350971ae00ea75b34caaa9db382f6)
Signed-off-by: Sebastiaan van Stijn <github@gone.nl>