api/types/network: split EndpointSettings config and operational data
EndpointSettings holds both "Config" data (definition) and runtime ("operational")
data. Fields for the latter one are set/reset at runtime. Resetting is handled
by the `cleanOperationalData` function, but it does so for each field individually.

This handling is slightly brittle, as the `cleanOperationalData` must be
updated when adding new "operational" data, which can easily be overlooked.

This patch updates the EndpointSettings type to split the "config" and
"operational" data into separate structs. The new `EndpointOperationalData`
struct is embedded in the existing type, so the net-result should be the
same for most cases (except for constructing struct-literals).

Resetting the operational data now allows for the `EndpointOperationalData`
to be set to an empty struct (resetting its values to their defaults / empty).

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>