integration: Make sure SYN probes are received by victim hosts
The tests introduced in "integration: Test port mapping security"
check victim hosts don't send a SYN-ACK when it should not to. This
could happen for two reasons: 1. because the right security measures
are effectively dropping the SYN packet ; or 2. because the SYN packet
was never received by the victim host.

The latter case could be a potential source of false positives, which
are pretty bad since we're testing security features. To alleviate this
risk, this commit adds the ability to capture packets on the victim's
host interface and leverage that to make sure the SYN packet is
well received. If it's not, the test fails.

Signed-off-by: Albin Kerouanton <albinker@gmail.com>