api: Endpoints can now return a tree of errors
There're cases like input validation where it makes sense to gather
multiple errors and return them all to the client.

Prior to this commit, the API would only return a string `message`
field. As a consequence, it requires those errors to be coalesced
into a single string using a specific format, whether it's just a
newline separator or some more complicated format.

This commit adds a new `errors` field containing a list of
ErrorResponse.

Signed-off-by: Albin Kerouanton <albinker@gmail.com>