volume/local: Unmount on startup only if needed
Previously, the unmount was performed only when loading the `opts.json`
file succeded. This protected from calling the unmount on Windows, but
also from unmounting _data directories NOT mounted by the daemon.
This is because local volume driver only performs the mount if it's
needed by the additional volume mount opts.

The behavior was changed when the mount was turned into a no-op on
Windows and the unmount was moved and became unconditional. This missed
the fact that the mount could NOT have been performed by the daemon, if
the volume didn't have any volume options.

This partially the old behavior and unmounts the data directory only if
conditions for mounting it are true.

Signed-off-by: Paweł Gronowski <pawel.gronowski@docker.com>